+++ config/celery.py
@@ -6,7 +6,7 @@ from celery import Celery
 # set the default Django settings module for the 'celery' program.
 os.environ.setdefault("DJANGO_SETTINGS_MODULE", "config.settings.local")
 
-app = Celery("example_site")
+app = Celery("kiosc")
 
 # Using a string here means the worker doesn't have to serialize
 # the configuration object to child processes.